+++ library.h
@@ -7,28 +7,28 @@ private:
 	std::string title;
 	std::string author;
 	std::string genre;
-	int page_count;			// page length 
-	int published;
+	unsigned short page_count;			// page length 
+	unsigned short published;
 	std::string reserved;		// checked out
 
 	// these should not be altered unless done direclty through the library constructor for newly archived books
 	void set_title(std::string title);
 	void set_author(std::string author);
 	void set_genre(std::string genre);
-	void set_page_count(int page_count);
-	void set_published(int published);
+	void set_page_count(unsigned short page_count);
+	void set_published(unsigned short published);
 
 public:
 	library();
 	library(std::string title, std::string author,
-		std::string genre, int page_count, int published, 
+		std::string genre, unsigned short page_count, unsigned short published,
 		std::string reserved);
 
 	std::string get_title() const;
 	std::string get_author() const;
 	std::string get_genre() const;
-	int get_page_count() const;
-	int get_published() const;
+	unsigned short get_page_count() const;
+	unsigned short get_published() const;
 	std::string get_reserved() const;
 
 	// reserved can be set true or false as needed
